% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Concatenate with Commas}
\usage{
cc(..., bracket = "")
}
\arguments{
\item{...}{zero or more objects which can be coerced to character}

\item{bracket}{A string to bracket the values by.}
}
\description{
Concatenates character vector(s) into a single string with each value separated
by a comma.
}
\examples{
cc(1:2,4)
cc(1,2,4, bracket = "'")
}
\seealso{
\code{\link{cc_or}} and \code{\link{cc_and}}
}
