% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc_and}
\alias{cc_and}
\title{Concatenation with Commas and And}
\usage{
cc_and(..., oxford = FALSE, bracket = "")
}
\arguments{
\item{...}{zero or more objects which can be coerced to character}

\item{oxford}{A flag indicating whether to used the Oxford comma}

\item{bracket}{A string to bracket the values by.}
}
\description{
Concatenates character vector(s) into a single string with each value separated
by a comma and the final value separated by and.
}
\examples{
cc_and(1:2,4)
cc_and(1,2,4, bracket = "'")
cc_and(1:4, oxford = TRUE)
}
\seealso{
\code{\link{cc}} and \code{\link{cc_or}}
}
