% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc-conjunction.R
\name{cc_conjunction}
\alias{cc_conjunction}
\alias{cc_and}
\alias{cc_or}
\title{Concatenation with Commas and Conjunction}
\usage{
cc_and(object, bracket = "'")

cc_or(object, bracket = "'")
}
\arguments{
\item{object}{The object with values to concatenate.}

\item{bracket}{A string to bracket the values by.}
}
\description{
Wrappers to concatenates object values with conjunction and by default brackets.
}
\section{Functions}{
\itemize{
\item \code{cc_and}: Concatenation with Commas and And Conjunction

\item \code{cc_or}: Concatenation with Commas and Or Conjunction
}}

\examples{
cc_and(1:100)
cc_or(100:1)
}
\seealso{
\code{\link{cc}}
}
