% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc_or}
\alias{cc_or}
\title{Concatenation with Commas and Or}
\usage{
cc_or(..., oxford = FALSE, bracket = "", ellipsis = 100)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character}

\item{oxford}{A flag indicating whether to used the Oxford comma}

\item{bracket}{A string to bracket the values by.}

\item{ellipsis}{A count of the total number of values required to begin using an ellipsis.}
}
\description{
Concatenates character vector(s) into a single string with each value separated
by a comma and the final value separated by or.
}
\examples{
cc_or(1:2,4)
cc_or(1,2,4, bracket = "'")
cc_or(1:4, oxford = TRUE)
}
\seealso{
\code{\link{cc}} and \code{\link{cc_and}}
}
