% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn.R
\name{cn}
\alias{cn}
\title{Number Aware Concatenation with Commas}
\usage{
cn(x, one, some = one, none = some, lots = none, bracket = "",
  nlots = 10)
}
\arguments{
\item{x}{A vector of length n that can be coerced to character.}

\item{one}{The string to return if n = 1}

\item{some}{The string to return if n is in 2, 3, ..., nlots - 1}

\item{none}{The string to return if n = 0}

\item{lots}{The string to return if n >= nlots}

\item{bracket}{A string to bracket the values by.}

\item{nlots}{A count of the number of values to consider to be lots.}
}
\description{
Concatenates character vector(s) into a single string with each value separated
by a comma.
}
\examples{
cn(1:3, "the \%n elements are: \%c")
fun <- function(x) {
 cn(x, 
  one = "there is 1 element: \%c",
  some = "there are \%n elements: \%c",
  none = "there are \%n elements",
  )
 }
 fun("3")
 fun(character(0))
 fun(3:2)
 fun(1:10)
}
\seealso{
\code{\link{cc}}, \code{\link{cn_or}} and \code{\link{cc_and}}
}
