% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn.R
\name{cn_and}
\alias{cn_and}
\title{Number Aware Concatenation with Commas and And}
\usage{
cn_and(object, one = "\%o has \%n value\%s: \%c", some = one,
  none = gsub(": ", "", some), lots = some, oxford = FALSE,
  bracket = "", ellipsis = 10, nlots = ellipsis)
}
\arguments{
\item{object}{An object that can be coerced to character vector of length n..}

\item{one}{The string to return if n = 1}

\item{some}{The string to return if n is in 2, 3, ..., nlots - 1}

\item{none}{The string to return if n = 0}

\item{lots}{The string to return if n >= nlots}

\item{oxford}{A flag indicating whether to used the Oxford comma}

\item{bracket}{A string to bracket the values by.}

\item{ellipsis}{A count of the total number of values required to begin using an ellipsis.}

\item{nlots}{A count of the number of values to consider to be lots.}
}
\description{
Number Aware Concatenation with Commas and And
}
\examples{
cn_and(character())
x <- "fox"
cn_and(x)
cn_and(c(1,2,5))
cn_and(1:10)
}
\seealso{
\code{\link{cc_and}}, \code{\link{cn}} and \code{\link{cc_or}}
}
