% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.R
\name{cs}
\alias{cs}
\title{Customisable S}
\usage{
cs(n, text = "\%s")
}
\arguments{
\item{n}{A count.}

\item{text}{A string of the message.}
}
\value{
A string of the updated message.
}
\description{
Customisable S
}
\examples{
cs(0)
cs(1)
cs(2)
cs(1, "value\%s")
cs(4, "value\%s")
}
\seealso{
\code{\link{cc}} and \code{\link{cn}}
}
