% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err-package.R
\docType{package}
\name{err-package}
\alias{err}
\alias{err-package}
\title{err: Messages}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

Authors:
\itemize{
  \item James Dunham \email{james@jamesdunham.io}
}

}
\keyword{internal}
