% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{err}
\alias{err}
\title{Error}
\usage{
err(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character}
}
\description{
Stops execution and throws an error without the call as part of the error message.
}
\seealso{
\code{\link[base]{stop}}, \code{\link{wrn}}, \code{\link{msg}} and \code{\link{co}}
}
