% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{msg}
\alias{msg}
\title{Message}
\usage{
msg(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character}
}
\description{
Generates a diagnostic message.
}
\seealso{
\code{\link[base]{message}}, \code{\link{err}}, \code{\link{wrn}} and and \code{\link{cn}}
}
