% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p}
\alias{p}
\alias{p0}
\title{Paste}
\usage{
p(..., sep = " ", collapse = NULL)

p0(..., collapse = NULL)
}
\arguments{
\item{...}{one or more \R objects, to be converted to character vectors.}

\item{sep}{a character string to separate the terms.  Not
    \code{\link[base]{NA_character_}}.}

\item{collapse}{an optional character string to separate the results.  Not
    \code{\link[base]{NA_character_}}.}
}
\value{
A string of the pasted values.
}
\description{
Wrappers on \code{\link{paste}} and \code{\link{paste0}} 
to increase the readability of code.
}
\section{Functions}{
\itemize{
\item \code{p0}: Paste0
}}

\examples{
p("The", "red")
p0("ard", "vark")
}
