% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err.R
\name{wrn}
\alias{wrn}
\title{Warning}
\usage{
wrn(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character}
}
\description{
Generates a warning message (without the call as part of the error message).
}
\seealso{
\code{\link[base]{warning}}, \code{\link{err}}, \code{\link{msg}} and \code{\link{co}}
}
