% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Concatenation with Commas}
\usage{
cc(..., bracket = "", ellipsis = 10)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character}

\item{bracket}{A string to bracket the values by.}

\item{ellipsis}{A count of the total number of values required to use an ellipsis.}
}
\description{
Concatenates character vector(s) into a single string with each value separated
by a comma.
}
\examples{
cc(1:2, 4)
cc(1, 2, 4, bracket = "'")
cc(1:100)
cc(1, 1, 1, 0, ellipsis = 4)
}
\seealso{
\code{\link{cc_or}} and \code{\link{cc_and}}
}
