% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc_and}
\alias{cc_and}
\title{Concatenation with Commas and And}
\usage{
cc_and(..., bracket = "'", ellipsis = 10, oxford = FALSE)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character}

\item{bracket}{A string to bracket the values by.}

\item{ellipsis}{A count of the total number of values required to use an ellipsis.}

\item{oxford}{A flag indicating whether to use the Oxford comma}
}
\description{
Concatenates character vector(s) into a single string with each value separated
by a comma and the final value separated by and.
}
\examples{
cc_and(1,1,1:2)
cc_and(100:1)
cc_and(100:1, bracket = "|", ellipsis = 5, oxford = TRUE)
}
\seealso{
\code{\link{cc}}, \code{\link{cc_or}} and \code{\link{co_and}}
}
