% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc_or}
\alias{cc_or}
\title{Concatenation with Commas and Or}
\usage{
cc_or(..., bracket = "'", ellipsis = 10, oxford = FALSE)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character}

\item{bracket}{A string to bracket the values by.}

\item{ellipsis}{A count of the total number of values required to use an ellipsis.}

\item{oxford}{A flag indicating whether to use the Oxford comma}
}
\description{
Concatenates character vector(s) into a single string with each value separated
by a comma and the final value separated by or.
}
\examples{
cc_or(1,1,1:2)
cc_or(100:1)
cc_or(100:1, bracket = "", ellipsis = 5, oxford = TRUE)
}
\seealso{
\code{\link{cc}}, \code{\link{cc_and}} and \code{\link{cn_or}}
}
