% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn.R
\name{cn}
\alias{cn}
\title{Number Aware Concatenation with Commas}
\usage{
cn(object, one = "\%o has \%n value\%s: \%c", some = one,
  none = gsub(": ", "", some), lots = some, bracket = "",
  ellipsis = 10, nlots = ellipsis)
}
\arguments{
\item{object}{An object that can be coerced to character vector of length n..}

\item{one}{The string to return if n = 1}

\item{some}{The string to return if n is in 2, 3, ..., nlots - 1}

\item{none}{The string to return if n = 0}

\item{lots}{The string to return if n >= nlots}

\item{bracket}{A string to bracket the values by.}

\item{ellipsis}{A count of the total number of values required to begin using an ellipsis.}

\item{nlots}{A count of the number of values to consider to be lots.}
}
\description{
Concatenates character vector(s) into a single string with each value separated
by a comma.
}
\examples{
cn(character())
x <- "fox"
cn(x)
cn(c(1,2,5))
cn(1:10)
}
\seealso{
\code{\link{cc}}, \code{\link{cn_or}} and \code{\link{cc_and}}
}
