% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn.R
\name{cn}
\alias{cn}
\title{Customisable Number Aware Message}
\usage{
cn(object, one = "\%o has \%n value\%s: \%c", some = one,
  none = gsub(": ", "", some), lots = some, bracket = "",
  ellipsis = 10, nlots = ellipsis)
}
\arguments{
\item{object}{An object that can be coerced to character vector of length n..}

\item{one}{The string to return if n = 1}

\item{some}{The string to return if n is in 2, 3, ..., nlots - 1}

\item{none}{The string to return if n = 0}

\item{lots}{The string to return if n >= nlots}

\item{bracket}{A string to bracket the values by.}

\item{ellipsis}{A count of the total number of values required to use an ellipsis.}

\item{nlots}{A count of the number of values to consider to be lots.}
}
\description{
Produces a fully customisable number aware message with consecutive values
separated by columns.
}
\section{\code{sprintf}-like types}{

The following \code{sprintf}-like types can be used in the custom messages:

\describe{
  \item{\code{c}}{the object as a comma separated list (producted by a \code{\link{cc}} function)}
  \item{\code{n}}{the length of the object}
  \item{\code{o}}{the name of the object}
  \item{\code{s}}{'s' if n != 1 otherwise ''}
}
}

\examples{
cn(character())
x <- "fox"
cn(x)
cn(c(1,2,5))
cn(1:10)
}
\seealso{
\code{\link{cn_or}}, \code{\link{cn_and}} and \code{\link{cc}}
}
