% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.R
\name{co}
\alias{co}
\alias{co_or}
\alias{co_and}
\title{Customisable Object Aware Message}
\usage{
co(object, one = "\%o has \%n value\%s: \%c", some = one,
  none = gsub(": ", "", some), lots = some, nlots = 10,
  conjunction = NULL, bracket = "'", ellipsis = nlots,
  oxford = FALSE, object_name = substitute(object), ...)

co_or(object, one = "\%o has \%n value\%s: \%c", some = one,
  none = gsub(": ", "", some), lots = some, nlots = 10,
  bracket = "'", ellipsis = nlots, oxford = FALSE,
  object_name = substitute(object), ...)

co_and(object, one = "\%o has \%n value\%s: \%c", some = one,
  none = gsub(": ", "", some), lots = some, nlots = 10,
  bracket = "'", ellipsis = nlots, oxford = FALSE,
  object_name = substitute(object), ...)
}
\arguments{
\item{object}{The object of length n.}

\item{one}{The string to return if n = 1}

\item{some}{The string to return if n is in 2, 3, ..., nlots - 1}

\item{none}{The string to return if n = 0}

\item{lots}{The string to return if n >= nlots}

\item{nlots}{A count of the number of values to consider to be lots.}

\item{conjunction}{A string of the conjunction to separate the last value by or NULL.}

\item{bracket}{A string to bracket the values by.}

\item{ellipsis}{A count of the total number of values required to use an ellipsis.}

\item{oxford}{A flag indicating whether to use the Oxford comma (if conjunction).}

\item{object_name}{A string of the object name.}

\item{...}{Unused.}
}
\description{
Produces a fully customisable object aware message with consecutive values
separated by columns.
}
\section{\code{sprintf}-like types}{

The following \code{sprintf}-like types can be used in the custom messages:

\describe{
  \item{\code{c}}{the object as a comma separated list (producted by a \code{\link{cc}} function)}
  \item{\code{n}}{the length of the object}
  \item{\code{o}}{the name of the object}
  \item{\code{s}}{'s' if n != 1 otherwise ''}
}
}

\examples{
co(character())
x <- "fox"
co(x)
co(c(1,2,5))
co(1:10)
co(datasets::mtcars)
}
\seealso{
\code{\link{cc}}
}
