% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrn}
\alias{wrn}
\title{Warning}
\usage{
wrn(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character}
}
\description{
Immediately outputs an warning without the call as part of the error message
that as far as possible is a single line.
}
\seealso{
\code{\link[base]{warning}}, \code{\link{err}}, \code{\link{msg}} and \code{\link{co}}
}
